修改了一些小问题

+++ 通用文件存储器项目/HHQDevelopmentKit/轮播图/3D轮播图/CustomLayout.m
@@ -35,7 +35,7 @@
     CGFloat arc = M_PI * 2.0f;
     
 
-    int numberOfVisibleItems = [self.collectionView numberOfItemsInSection:0 ];
+    NSInteger numberOfVisibleItems = [self.collectionView numberOfItemsInSection:0 ];
     
     attributes.center = CGPointMake(x + collection.frame.size.width/2 , collection.frame.size.height/2);
